lambdas/pyproject.toml
[tool.poetry]
name = "lambdas"
version = "0.1.0"
description = ""
authors = ["Anthony Viera <anthonyviera4@gmail.com>"]

packages = [
    { include = "src" }
]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "0.115.3"
mangum = "^0.17.0"
boto3 = "^1.34.55"
aws-lambda-powertools = "2.22.0"
uuid = "^1.30"
black = "^24.2.0"
arrow = "^1.3.0"
pydantic = "^2.6.4"
pytest = "^7.1.2"
requests-mock = "^1.11.0"
freezegun = "^1.2.1"
mock = "^5.1.0"
simplejson = "^3.19.2"
pytest-cov = "^4.1.0"
pytest-helpers-namespace = "^2021.12.29"
urllib3 = "^1.26.19"
python-multipart = "0.0.18"
pytest-mock = "^3.14.0"
aws-lambda-context = "^1.1.0"
requests = "^2.32.0"
idna = "^3.7"
uvicorn = "^0.32.1"
pyyaml = "^6.0.2"
isort = "^5.13.2"

[tool.poetry.dev-dependencies]

[tool.poetry.group.dev.dependencies]
httpx = "^0.27.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.coverage.run]
branch = false
source = ["src"]
omit = [
    "src/api.py",
]

[tool.coverage.paths]
source = ["src"]

[tool.coverage.report]
show_missing = true
fail_under = 100

[tool.poetry.scripts]
generate_swagger = "generate_swagger:main"